additional notes about the issue
The only possible values of getRoomX and getRoomZ are either -8 and 0,
or 0 and 8, due to the forloop the calls them
(MazeRoomBlockPopulator@L30). Seems that the negative case was never
tested for and never discovered as CB never complained about it.
Therefore, all of the rest of the computations will never exceed 15, so
the other problem (coordinate too large) never occurs.